feat: refactor file structure. this commit resolves #3
