--- src/components/layout/index.js
@@ -1,8 +1,0 @@
-// Layouts
-import Header from './Header'
-import Footer from './Footer'
-
-export {
-  Header,
-  Footer
-}
