+++ src/containers/DominicansWhoCodesList.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
 import Row from 'reactstrap/lib/Row'
-import CoderCard from '../../components/CoderCard'
+import CoderCard from '../components/CoderCard'
 
-import useDominicanCoders from '../../hooks/useDominicanCoders'
+import { useDominicanCoders } from '../hooks/useDominicanCoders'
 
 const DominicansWhoCodesList = () => {
     const { data, loading } = useDominicanCoders()
