+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './sass/index.scss'
-import App from './App'
+import { App } from './containers/App'
 
 ReactDOM.render(<App />, document.getElementById('root'))
